feat : 타입 정의
