+++ frontend/src/pages/HomePage/HomePage.tsx
@@ -1,8 +1,15 @@
 import React, { useState, useEffect } from 'react';
 import WebtoonService from '@/services/webtoonService';
-import { HomePageState } from '@/types/page';
+import { Webtoon } from '@/types/webtoon';  
 import WebtoonGrid from '@/components/WebtoonGrid';
 import styles from './HomePage.module.css';
+
+export interface HomePageState {
+  popularWebtoons: Webtoon[];
+  recentWebtoons: Webtoon[];
+  isLoading: boolean;
+  error: string | null;
+}
 
 const HomePage: React.FC = () => {
   const [state, setState] = useState<HomePageState>({
@@ -11,7 +18,7 @@ const HomePage: React.FC = () => {
     isLoading: true,
     error: null
   });
-/*
+
   useEffect(() => {
     const fetchData = async () => {
       try {
@@ -37,7 +44,7 @@ const HomePage: React.FC = () => {
 
     fetchData();
   }, []);
-  */
+  
  
   if (state.isLoading) return <div>로딩중...</div>;
   if (state.error) return <div>{state.error}</div>;
@@ -47,13 +54,13 @@ const HomePage: React.FC = () => {
       {/* 인기 웹툰 섹션 */}
       <section className={styles.section}>
         <h2>인기 웹툰</h2>
-        {/* <WebtoonGrid webtoons={state.popularWebtoons || []} /> */}
+         <WebtoonGrid webtoons={state.popularWebtoons || []} /> 
       </section>
 
       {/* 최신 웹툰 섹션 */}
       <section className={styles.section}>
         <h2>최신 웹툰</h2>
-        {/* <WebtoonGrid webtoons={state.recentWebtoons || []} /> */}
+         <WebtoonGrid webtoons={state.recentWebtoons || []} /> 
       </section>
     </div>
   );
